bring back update option
I like having the simple `update` option for people who do not want to
get the remote file and merge in separate steps.

I renamed `update-remote` to `fetch-updates` in the hopes that the
difference between `update` and `fetch-updates` will not be too
confusing (we'll explain the difference in the docs too, obviously).

I also tweaked the message displayed after downloading the file. After
the new file is downloaded, we're showing the user the total count of
entries. We're not diffing it against a previous version or anything, so
they are not necesarilly "new" entries.