de-dupe to prevent remote hosts from overriding enabled/disabled
This ended up being slightly trickier than I thought. There may be a
more elegant way to go about it.

We cat all the files together except the remote hosts. We strip out all
comment lines from remote hosts to make our lives easier (they don't do
us any good anyway -- if a user were to inspect or modify the file they
should do it in /etc/hostsctl.d, not /etc/hosts -- and, hey, we save a
few bytes in the result). We then remove from remote hosts any line
which matches an entry in the enabled-disabled file. To do this we're
actually searching against the lines without a leading `#` (ie, we're
searching as if all the hosts in the file were disabled). Then we append
the de-duped remote hosts file to the chunk we had previously
concatenated.

The result is basically the same as before, except that now the remote
hosts file can not override any hosts that have been explicitly enabled
or disabled via hostsctl.

closes #4