avoid gnu specific awk flag
-v is only available in GNU awk, not the one true awk.

The awk statement in hosts_list() should behave exactly the same.

The awk statement in hosts_fetch_updates() has changed. Previously
entries beginning with "#0.0" were allowed, but those are useless, so
they are now stripped with the rest of the comments in the remote hosts
file. This should have no noticeable impact for the user