count modified entries, not "new" entries
We don't actually have any idea which entries are new. By counting the
number of lines that differ between the old and new remote hosts file we
can provide the user with a useful metric on their up-to-date-ness,
without misleading them with terms like "new entries".

This change also avoids a bug in the old implementation that occurs in
whatever bash Apple ships.